fix(doctor): reject incomplete exec approvals migration
Check canonical approvals readiness after migration contributions and before
managed maintenance can restart the Gateway. Preserve legacy bytes and
existing policy while reporting incomplete repair as an error.

Fixes #135437.
